# back-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Juegos",
        "description": "API para la gestión de juegos, rondas y participantes. Permite crear juegos, gestionar rondas y administrar participantes con sus respectivos resultados.",
        "contact": {
            "name": "Soporte API",
            "email": "fabriciocervantesmendoza@gmail.com"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://opensource.org/licenses/AGPL-3.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
        "/api/game": {
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Crear un nuevo juego",
                "description": "Crea un nuevo juego con descripción y monto. La fecha de juego se establece automáticamente.",
                "operationId": "7a2506f2777e2d3f3d3e2290e4635687",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "descrip",
                                    "monto"
                                ],
                                "properties": {
                                    "descrip": {
                                        "description": "Descripción del juego",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Torneo de Ajedrez Mensual"
                                    },
                                    "monto": {
                                        "description": "Monto del juego",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Juego creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "game": {
                                            "$ref": "#/components/schemas/Game"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "descrip": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The descrip field is required."
                                                    }
                                                },
                                                "monto": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The monto must be at least 0."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/game/cierre": {
            "put": {
                "tags": [
                    "Games"
                ],
                "summary": "Actualizar fecha de cierre del juego",
                "description": "Establece la fecha de cierre de un juego. La fecha debe ser posterior al momento actual.",
                "operationId": "d99de8bb00c7172defa29b65089458fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fec_cierre"
                                ],
                                "properties": {
                                    "fec_cierre": {
                                        "description": "Fecha y hora de cierre del juego (debe ser futura)",
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2024-08-10 18:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Juego actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Juego actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Game"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Juego no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Game]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "fec_cierre": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The fec cierre must be a date after now."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Game": {
                "title": "Juego",
                "description": "Modelo de juego",
                "required": [
                    "id",
                    "descrip",
                    "monto",
                    "fec_juego"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del juego",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "descrip": {
                        "description": "Descripción del juego",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Torneo de Ajedrez Mensual"
                    },
                    "monto": {
                        "description": "Monto del juego",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 50
                    },
                    "fec_juego": {
                        "description": "Fecha del juego",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-05T10:00:00Z"
                    },
                    "fec_cierre": {
                        "description": "Fecha de cierre del juego",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-10T18:00:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-01T09:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-02T11:15:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT en el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}